+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- fix: Funscript generation with Stash videos was using the internal video ID instead of the Stash scene ID.
 
 ## 0.17.0
 - feat: Unified video library. You no longer need to choose between whether you want videos from Stash or from your local hard drive.